refactor(users): update imports
