+++ src/users/application/update-user.use-case.ts
@@ -1,13 +1,16 @@
 import { Inject, Injectable } from '@nestjs/common'
 import { UsersTokens } from 'users/di/users.tokens'
 import { UsersRepository } from 'users/domain/users.repository'
-import { UpdateUserDto } from 'users/domain/dto'
+import { UpdateUserDto, UsersDtoEntityMapper } from 'users/application/dto'
 
 @Injectable()
 export class UpdateUser {
-	constructor(@Inject(UsersTokens.USER_REPOSITORY) private readonly repository: UsersRepository) {}
+	constructor(
+		@Inject(UsersTokens.USER_REPOSITORY) private readonly repository: UsersRepository,
+		private readonly mapper: UsersDtoEntityMapper
+	) {}
 
 	run(id: string, data: UpdateUserDto) {
-		return this.repository.updateUser(id, data)
+		return this.repository.updateUser(id, this.mapper.toEntity(data))
 	}
 }
